refactor(menubar): satisfy layer, size, and POSIX architecture guards
- Frozen-launch helpers move to quodeq/shared/frozen.py (stdlib-only,
  cross-cutting) with a dashboard/_frozen re-export shim, so menubar
  never imports the dashboard layer; new menubar layer rule added.
- Dashboard lifecycle methods split into _app_lifecycle.py mixin to
  keep app.py under the 300-line file cap.
- lsof/killpg/pkill sites allowlisted: the package is darwin-only by
  construction and each call degrades gracefully.
