+++ src/quodeq/menubar/app.py
@@ -8,47 +8,36 @@ window's. Differences from the retired QuodeqBar app:
 - Launches the dashboard by re-executing its own binary (dashboard_cmd)
   instead of discovering a ``quodeq`` CLI on PATH.
 - Quit flips the Settings preference off so the icon stays gone.
+
+Dashboard start/stop/open lives in _app_lifecycle.py (mixin, split for size).
 """
 from __future__ import annotations
 
 import logging as _logging
 import os
-import signal
 import subprocess
 import threading
 
 import rumps
 
-from quodeq.dashboard._frozen import dashboard_cmd as _dashboard_cmd
-from quodeq.dashboard._frozen import source_user_path as _source_user_path
+from quodeq.shared.frozen import source_user_path as _source_user_path
 from quodeq.menubar import control as _control
 from quodeq.menubar import state as _state
+from quodeq.menubar._app_lifecycle import DashboardLifecycleMixin
 from quodeq.menubar._health import (
     find_commands as _find_commands,
     find_icon as _find_icon,
     health_check as _health_check,
     is_evaluating as _is_evaluating,
 )
-from quodeq.menubar._process import (
-    DashboardCallbacks as _DashboardCallbacks,
-    DashboardState as _DashboardState,
-    cleanup_stderr_log as _cleanup_stderr_log_file,
-    find_running_port as _find_running_port_cached,
-    kill_port_processes as _kill_port_processes,
-    open_stderr_log as _open_stderr_log,
-    wait_for_dashboard as _wait_for_dashboard,
-    _ERROR_DISPLAY_MAX,
-    _STDERR_READ_MAX,
-)
+from quodeq.menubar._process import find_running_port as _find_running_port_cached
 
-_PKILL_TIMEOUT_S = 5
 _DEFAULT_APP_PORT = 7863
 try:
     _POLL_INTERVAL = int(os.environ.get("QUODEQ_POLL_INTERVAL", "5"))
 except ValueError:
     _logging.getLogger(__name__).warning("Invalid QUODEQ_POLL_INTERVAL; using default 5")
     _POLL_INTERVAL = 5
-_PROCESS_PATTERNS = ("quodeq.api.app", "quodeq.action_api", "quodeq dashboard")
 _DEFAULT_PORTS = "7863,7864,7865,7866,7867,7868,7869"
 
 
@@ -72,7 +61,7 @@ def _load_config(env=None):
     return app_port, ports
 
 
-class QuodeqApp(rumps.App):
+class QuodeqApp(DashboardLifecycleMixin, rumps.App):
     def __init__(self):
         super().__init__(
             "Quodeq", icon=_find_icon("menubar_iconTemplate.png"), template=True,
@@ -213,137 +202,6 @@ class QuodeqApp(rumps.App):
                 self._update_item.title = "Check for Updates…"
         except Exception:
             pass
-
-    def _on_open(self, _):
-        with self._state_lock:
-            port = self._port
-        port = port or self._find_running_port()
-        if not port:
-            return
-        # Launches the dashboard app (or brings the existing window to the
-        # front via its socket IPC) by re-executing our own binary.
-        subprocess.Popen(
-            _dashboard_cmd(["--no-build", "--port", str(port)]),
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            start_new_session=True,
-        )
-
-    def _on_start(self, _):
-        threading.Thread(target=self._do_start, daemon=True).start()
-
-    def _do_start(self):
-        with self._state_lock:
-            if self._starting:
-                return
-            if self._find_running_port():
-                return
-            self._starting = True
-        self._clear_error()
-        try:
-            self._do_start_inner()
-        except (OSError, subprocess.SubprocessError, ValueError) as e:
-            self._set_error(f"Error: {e}")
-            self._status_item.title = "Stopped"
-            self._cleanup_stderr_log()
-        finally:
-            with self._state_lock:
-                self._starting = False
-
-    def _cleanup_stderr_log(self) -> None:
-        """Remove the stderr log tempfile if it exists."""
-        _cleanup_stderr_log_file(self._stderr_log_path)
-        self._stderr_log_path = None
-
-    def _launch_dashboard_process(self, stderr_log) -> bool:
-        """Launch the dashboard subprocess. Returns True on success, False on failure."""
-        try:
-            cmd = _dashboard_cmd(["--no-open", "--port", str(self._app_port)])
-            self._process = subprocess.Popen(
-                cmd, stdout=subprocess.DEVNULL, stderr=stderr_log, start_new_session=True,
-            )
-            return True
-        except OSError as e:
-            stderr_log.close()
-            self._set_error(f"Failed: {e}")
-            self._status_item.title = "Stopped"
-            self._cleanup_stderr_log()
-            return False
-
-    def _do_start_inner(self):
-        self._status_item.title = "Starting..."
-        stderr_log = _open_stderr_log()
-        self._stderr_log_path = stderr_log.name
-        if not self._launch_dashboard_process(stderr_log):
-            return
-        self._wait_for_dashboard(stderr_log)
-
-    def _handle_crashed_process(self, stderr_log) -> None:
-        """Report a crashed dashboard process and clean up."""
-        stderr_log.close()
-        try:
-            with open(stderr_log.name, encoding="utf-8") as f:
-                err = f.read(_STDERR_READ_MAX).strip()
-        except OSError:
-            err = "unknown error"
-        sanitized = err[:_ERROR_DISPLAY_MAX].replace("\n", " ").strip()
-        if sanitized:
-            # Keep the crash detail in the local log for troubleshooting, but
-            # do not surface raw dashboard stderr (which may include tokens or
-            # filesystem paths) in the always-visible menu.
-            _logging.getLogger(__name__).warning(
-                "Dashboard crashed (exit code %s): %s", self._process.returncode, sanitized,
-            )
-        self._set_error(
-            f"Dashboard stopped unexpectedly (exit code {self._process.returncode}). Try restarting."
-        )
-        self._status_item.title = "Stopped"
-        self._cleanup_stderr_log()
-
-    def _wait_for_dashboard(self, stderr_log):
-        """Poll until the dashboard responds or process crashes."""
-        def on_port_found(port, _log):
-            with self._state_lock:
-                self._port = port
-            self._clear_error()
-            self._cleanup_stderr_log()
-
-        def on_timeout():
-            self._set_error("Timeout: dashboard did not respond")
-            self._status_item.title = "Stopped"
-            self._cleanup_stderr_log()
-
-        _wait_for_dashboard(
-            process=self._process,
-            ports=self._ports,
-            state=_DashboardState(cache=self._port_cache, last_known=self._port_cache.get("last_known")),
-            stderr_log=stderr_log,
-            callbacks=_DashboardCallbacks(
-                on_port_found=on_port_found,
-                on_crash=self._handle_crashed_process,
-                on_timeout=on_timeout,
-            ),
-        )
-
-    def _on_stop(self, _):
-        if self._process and self._process.poll() is None:
-            try:
-                os.killpg(os.getpgid(self._process.pid), signal.SIGTERM)
-            except (OSError, ProcessLookupError):
-                self._process.terminate()
-            self._process = None
-        for port in self._ports:
-            _kill_port_processes(port)
-        for pattern in _PROCESS_PATTERNS:
-            try:
-                subprocess.run(["pkill", "-f", pattern], capture_output=True, timeout=_PKILL_TIMEOUT_S)
-            except (subprocess.TimeoutExpired, OSError):
-                pass
-        with self._state_lock:
-            self._port = None
-        self._status_item.title = "Stopped"
-        self._set_ui_state(running=False)
-        self._cleanup_stderr_log()
 
 
 def _set_accessory_policy() -> None:
